refactor code indexer
